+++ src/main/java/be/cytomine/service/project/ProjectService.java
@@ -503,7 +503,7 @@ public class ProjectService extends ModelService {
         for (Map.Entry<String, Object> entry : mapParams.entrySet()) {
             query.setParameter(entry.getKey(), entry.getValue());
         }
-        long count = ((BigInteger)query.getResultList().get(0)).longValue();
+        long count = (Long)query.getResultList().get(0);
         Page<JsonObject> page = PageUtils.buildPageFromPageResults(results, max, offset, count);
         return page;
